From dynamic to static lib

+++ Brynhild/src/Brynhild/ImGui/ImGuiLayer.cpp
@@ -68,7 +68,7 @@ namespace Brynhild {
     ImGuiIO& IO = ImGui::GetIO();
 
     Application& app = Application::Get();
-    IO.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+    IO.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
     ImGui::Render();
     ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
